id as long

+++ src/main/java/org/yorm/Yorm.java
@@ -23,10 +23,10 @@ public class Yorm {
         this.queryBuilder = new QueryBuilder(ds);
     }
 
-    public int save(Record recordObj) throws YormException {
+    public long save(Record recordObj) throws YormException {
         String objectName = getRecordName(recordObj);
         YormTable yormTable = getTable(objectName, recordObj.getClass());
-        int result = 0;
+        long result = 0;
         try {
             result = queryBuilder.save(ds, recordObj, yormTable);
         } catch (InvocationTargetException | IllegalAccessException e) {
@@ -35,7 +35,7 @@ public class Yorm {
         return result;
     }
 
-    public int insert(Record recordObj) throws YormException {
+    public long insert(Record recordObj) throws YormException {
         String objectName = getRecordName(recordObj);
         YormTable yormTable = getTable(objectName, recordObj.getClass());
         return queryBuilder.insert(ds, recordObj, yormTable);
@@ -58,7 +58,7 @@ public class Yorm {
         queryBuilder.update(ds, recordObj, yormTable);
     }
 
-    public <T extends Record> T find(Class<T> recordObject, int id) throws YormException {
+    public <T extends Record> T find(Class<T> recordObject, long id) throws YormException {
         String objectName = getClassName(recordObject);
         YormTable yormTable = getTable(objectName, recordObject);
         return queryBuilder.get(ds, yormTable, id);
@@ -108,7 +108,7 @@ public class Yorm {
         return recordObj.getClass().getSimpleName().toLowerCase(Locale.ROOT);
     }
 
-    public <T extends Record> boolean delete(Class<T> recordObject, int id) throws YormException {
+    public <T extends Record> boolean delete(Class<T> recordObject, long id) throws YormException {
         String objectName = getClassName(recordObject);
         YormTable yormTable = getTable(objectName, recordObject);
         return queryBuilder.delete(ds, yormTable, id);
