+++ src/main/java/org/yorm/db/operations/QueryDelete.java
@@ -12,14 +12,14 @@ public class QueryDelete {
     private QueryDelete() {
     }
 
-    public static boolean delete(DataSource ds, YormTable yormTable, int id) throws YormException {
+    public static boolean delete(DataSource ds, YormTable yormTable, long id) throws YormException {
         StringBuilder query = new StringBuilder("DELETE ");
         query.append(" FROM ")
             .append(yormTable.dbTable())
             .append(" WHERE id=?");
         try (Connection connection = ds.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(query.toString())) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             return preparedStatement.executeUpdate() > 0;
         } catch (SQLException e) {
             throw new YormException("Error while deleting record with id:" + id + " from table:" + yormTable.dbTable(), e);
