+++ src/main/java/org/yorm/db/operations/QueryFind.java
@@ -39,13 +39,13 @@ public class QueryFind {
         return resultList;
     }
 
-    public static <T extends Record> T findById(DataSource ds, YormTable yormTable, int id) throws YormException {
+    public static <T extends Record> T findById(DataSource ds, YormTable yormTable, long id) throws YormException {
         List<YormTuple> tuples = yormTable.tuples();
         String query = yormTable.selectAllFromTable() + " WHERE ID = ?";
         Object result = null;
         try (Connection connection = ds.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             ResultSet rs = preparedStatement.executeQuery();
             if (rs.next()) {
                 Object[] values = new Object[tuples.size()];
@@ -61,13 +61,13 @@ public class QueryFind {
         return (T) result;
     }
 
-    public static <T extends Record> List<T> findByForeignId(DataSource ds, YormTable yormTable, String fieldName, int id) throws YormException {
+    public static <T extends Record> List<T> findByForeignId(DataSource ds, YormTable yormTable, String fieldName, long id) throws YormException {
         List<T> resultList = new ArrayList<>();
         List<YormTuple> tuples = yormTable.tuples();
         String query = yormTable.selectAllFromTable() + " WHERE " + fieldName + " = ?";
         try (Connection connection = ds.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(query.toString())) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             ResultSet rs = preparedStatement.executeQuery();
             while (rs.next()) {
                 Object[] values = new Object[tuples.size()];
