+++ src/test/java/org/yorm/records/Person.java
@@ -2,6 +2,6 @@ package org.yorm.records;
 
 import java.time.LocalDateTime;
 
-public record Person(int id, String name, String email, LocalDateTime lastLogin, int companyId) {
+public record Person(long id, String name, String email, LocalDateTime lastLogin, int companyId) {
 
 }
